下载优化

+++ src/MonsterSiren.Uwp/Services/DownloadService.cs
@@ -1,5 +1,6 @@
 using System.Collections.ObjectModel;
 using System.Net.Http;
+using System.Text;
 using System.Text.Json;
 using System.Threading;
 using Windows.Media.Core;
@@ -23,10 +24,10 @@ public static class DownloadService
     private static bool _downloadLyric = true;
     private static bool _transcodeDownloadedItem = true;
     private static bool _replaceInvalidCharInFileName = true;
-    [Obsolete("Use new settings instead")]
-    private static CodecInfo _transcodeEncoderInfo;
+    private static bool _allowUnnecessaryTranscode = true;
     private static AudioFormat _transcodeFormat;
     private static AudioEncodingQuality _transcodeQuality = AudioEncodingQuality.High;
+    private static string _musicFileTemplateString;
     private static readonly BackgroundDownloader Downloader = new()
     {
         CostPolicy = BackgroundTransferCostPolicy.Always
@@ -69,20 +70,6 @@ public static class DownloadService
         {
             SettingsHelper.Set(CommonValues.MusicTranscodeDownloadedItemSettingsKey, value);
             _transcodeDownloadedItem = value;
-        }
-    }
-
-    [Obsolete("Use new settings instead")]
-    /// <summary>
-    /// 获取或设置转码操作要使用的编码器信息。
-    /// </summary>
-    public static CodecInfo TranscodeEncoderInfo
-    {
-        get => _transcodeEncoderInfo;
-        set
-        {
-            SettingsHelper.Set(CommonValues.MusicTranscodeEncoderGuidSettingsKey, value.Subtypes[0]);
-            _transcodeEncoderInfo = value;
         }
     }
 
@@ -138,6 +125,31 @@ public static class DownloadService
         }
     }
 
+    /// <summary>
+    /// 指示是否允许不必要的转码。
+    /// </summary>
+    public static bool AllowUnnecessaryTranscode
+    {
+        get => _allowUnnecessaryTranscode;
+        set
+        {
+            SettingsHelper.Set(CommonValues.MusicAllowUnnecessaryTranscodeSettingsKey, value);
+            _allowUnnecessaryTranscode = value;
+        }
+    }
+
+    /// <summary>
+    /// 获取或更改音乐文件的文件名模板。
+    /// </summary>
+    public static string MusicFileTemplateString
+    {
+        get => _musicFileTemplateString;
+        set
+        {
+            SettingsHelper.Set(CommonValues.MusicFileTemplateStringSettingsKey, value);
+            _musicFileTemplateString = value;
+        }
+    }
 
     /// <summary>
     /// 获取下载列表。
@@ -182,6 +194,11 @@ public static class DownloadService
                                     && keepWav;
         }
 
+        {
+            AllowUnnecessaryTranscode = SettingsHelper.TryGet(CommonValues.MusicAllowUnnecessaryTranscodeSettingsKey, out bool unnecessaryTranscode)
+                                    && unnecessaryTranscode;
+        }
+
         if (SettingsHelper.TryGet(CommonValues.MusicTranscodeFormatSettingsKey, out string formatString) && Enum.TryParse(formatString, out AudioFormat format))
         {
             TranscodeFormat = format;
@@ -203,6 +220,15 @@ public static class DownloadService
         else
         {
             TranscodeFormat = AudioFormat.Mp3;
+        }
+
+        if (SettingsHelper.TryGet(CommonValues.MusicFileTemplateStringSettingsKey, out string template) && !string.IsNullOrWhiteSpace(template))
+        {
+            MusicFileTemplateString = template;
+        }
+        else
+        {
+            MusicFileTemplateString = CommonValues.DefaultMusicFilenameTemplate;
         }
 
         if (EnvironmentHelper.IsWindowsMobile)
@@ -287,13 +313,38 @@ public static class DownloadService
 
         await Task.Run(async () =>
         {
+            string defaultMsrName = "MSR".GetLocalized();
             string rawMusicExtensions = Path.GetExtension(songDetail.SourceUrl) ?? ".wav";
-            string musicName = songDetail.Name?.Trim();
-            string musicFileName = $"{songDetail.Artists.FirstOrDefault()?.Trim() ?? "MSR".GetLocalized()} - {musicName}".Trim();
+            CreationCollisionOption collisionOption = CreationCollisionOption.ReplaceExisting;
+
+            string albumTitle = albumDetail.Name?.Trim();
+            string songTitle = songDetail.Name?.Trim();
+            string artist = songDetail.Artists.FirstOrDefault()?.Trim() ?? defaultMsrName;
+            string artists = string.Join(',', songDetail.Artists);
+            if (string.IsNullOrWhiteSpace(artists))
+            {
+                artists = defaultMsrName;
+            }
+
+            StringBuilder musicFileNameBuilder = new(MusicFileTemplateString);
+            foreach (string template in CommonValues.MusicFilenamePartTemplates)
+            {
+                string content = template switch
+                {
+                    "{AlbumTitle}" => albumTitle,
+                    "{SongTitle}" => songTitle,
+                    "{Artist}" => artist,
+                    "{Artists}" => artists,
+                    _ => throw new NotImplementedException("未添加对指定文件名模板的支持。")
+                };
+                musicFileNameBuilder.Replace(template, content);
+            }
+
+            string musicFileName = musicFileNameBuilder.ToString();
 
             if (ReplaceInvalidCharInFileName)
             {
-                musicFileName = CommonValues.ReplaceInvaildFileNameChars(musicFileName);
+                musicFileName = CommonValues.ReplaceInvalidFileNameChars(musicFileName);
             }
             else
             {
@@ -314,14 +365,14 @@ public static class DownloadService
 
             if (targetItem is not null && targetItem.IsOfType(StorageItemTypes.File) && (await targetItem.GetBasicPropertiesAsync()).Size != 0)
             {
-                DownloadItem item = new(musicName);
+                DownloadItem item = new(songTitle);
                 await AddToList(item);
                 return;
             }
 
-            StorageFile musicFile = await albumFolder.CreateFileAsync($"{musicFileName}{rawMusicExtensions}.tmp", CreationCollisionOption.ReplaceExisting);
+            StorageFile musicFile = await albumFolder.CreateFileAsync($"{musicFileName}{rawMusicExtensions}.tmp", collisionOption);
 
-            StorageFile infoFile = await albumFolder.CreateFileAsync($"{musicFileName}.json.tmp", CreationCollisionOption.ReplaceExisting);
+            StorageFile infoFile = await albumFolder.CreateFileAsync($"{musicFileName}.json.tmp", collisionOption);
             SongDetailAndAlbumDetailPack pack = new(songDetail, albumDetail);
             Stream infoFileStream = await infoFile.OpenStreamForWriteAsync();
             infoFileStream.Seek(0, SeekOrigin.Begin);
@@ -329,13 +380,13 @@ public static class DownloadService
             infoFileStream.Dispose();
 
             DownloadOperation musicDownload = Downloader.CreateDownload(new Uri(songDetail.SourceUrl, UriKind.Absolute), musicFile);
-            bool isSuccess = await HandleDownloadOperation(musicDownload, musicName, true);
+            bool isSuccess = await HandleDownloadOperation(musicDownload, songTitle, true);
 
             if (isSuccess && DownloadLyric && Uri.TryCreate(songDetail.LyricUrl, UriKind.Absolute, out Uri lrcUri))
             {
-                StorageFile lrcFile = await albumFolder.CreateFileAsync($"{musicFileName}.lrc.tmp", CreationCollisionOption.ReplaceExisting);
+                StorageFile lrcFile = await albumFolder.CreateFileAsync($"{musicFileName}.lrc.tmp", collisionOption);
                 DownloadOperation lrcDownload = Downloader.CreateDownload(lrcUri, lrcFile);
-                await HandleDownloadOperation(lrcDownload, $"{musicName} - {"LyricFile".GetLocalized()}", true);
+                await HandleDownloadOperation(lrcDownload, $"{songTitle} - {"LyricFile".GetLocalized()}", true);
             }
         });
     }
@@ -432,26 +483,38 @@ public static class DownloadService
                 dlItem.Progress = 0d;
                 dlItem.State = DownloadItemState.Transcoding;
 
-                StorageFolder destinationFolder = await sourceFile.GetParentAsync();
-                string destinationFileExtensions = $".{TranscodeFormat.ToString().ToLower()}";
-                string sourceFileExtension = Path.GetExtension(sourceFile.Path);
+                StorageFile destinationFile;
 
-                string destinationFileName = Path.ChangeExtension(sourceFile.Name, destinationFileExtensions);
-
-                if (destinationFileExtensions.Equals(sourceFileExtension, StringComparison.OrdinalIgnoreCase))
+                if (sourceFile.ContentType != "audio/mpeg"
+                    || TranscodeFormat == AudioFormat.Mp3
+                    || AllowUnnecessaryTranscode)
                 {
-                    await sourceFile.RenameAsync($"{sourceFile.DisplayName}{"RawMusicFileSaveTag".GetLocalized()}{sourceFileExtension}", NameCollisionOption.ReplaceExisting);
-                }
-                
-                StorageFile destinationFile = await destinationFolder.CreateFileAsync(destinationFileName, CreationCollisionOption.ReplaceExisting);
+                    StorageFolder destinationFolder = await sourceFile.GetParentAsync();
 
-                await TranscodeFile(sourceFile, destinationFile, TranscodeFormat, TranscodeQuality, dlItem);
+                    string destinationFileExtensions = $".{TranscodeFormat.ToString().ToLower()}";
+                    string destinationFileName = Path.ChangeExtension(sourceFile.Name, destinationFileExtensions);
+
+                    string sourceFileExtension = Path.GetExtension(sourceFile.Path);
+
+                    if (destinationFileExtensions.Equals(sourceFileExtension, StringComparison.OrdinalIgnoreCase))
+                    {
+                        await sourceFile.RenameAsync($"{sourceFile.DisplayName}{"RawMusicFileSaveTag".GetLocalized()}{sourceFileExtension}", NameCollisionOption.ReplaceExisting);
+                    }
+
+                    destinationFile = await destinationFolder.CreateFileAsync(destinationFileName, CreationCollisionOption.ReplaceExisting);
+                    await TranscodeFile(sourceFile, destinationFile, TranscodeFormat, TranscodeQuality, dlItem);
+
+                    if (KeepRawMusicFileAfterTranscode != true)
+                    {
+                        await sourceFile.DeleteAsync(StorageDeleteOption.PermanentDelete);
+                    }
+                }
+                else
+                {
+                    destinationFile = sourceFile;
+                }
+
                 await WriteTagsToFile(destinationFile, dlItem);
-
-                if (KeepRawMusicFileAfterTranscode != true)
-                {
-                    await sourceFile.DeleteAsync(StorageDeleteOption.PermanentDelete);
-                }
             }
             else
             {
@@ -460,40 +523,6 @@ public static class DownloadService
         }
 
         dlItem.State = DownloadItemState.Done;
-    }
-
-    [Obsolete("Don't use")]
-    private static MediaEncodingProfile GetEncodingProfile()
-    {
-        if (TranscodeEncoderInfo is null)
-        {
-            throw new InvalidOperationException($"{TranscodeEncoderInfo} 为 null");
-        }
-        else if (TranscodeEncoderInfo.Kind != CodecKind.Audio || TranscodeEncoderInfo.Category != CodecCategory.Encoder)
-        {
-            throw new InvalidOperationException($"{TranscodeEncoderInfo} 不是音频编码器");
-        }
-        else if (TranscodeQuality == AudioEncodingQuality.Auto)
-        {
-            throw new InvalidOperationException($"{TranscodeQuality} 被设为 '{AudioEncodingQuality.Auto}'");
-        }
-
-        IReadOnlyList<string> subtypes = TranscodeEncoderInfo.Subtypes;
-        if (subtypes.Any(IsTargetEncoder(CodecSubtypes.AudioFormatMP3)))
-        {
-            return MediaEncodingProfile.CreateMp3(TranscodeQuality);
-        }
-        else if (subtypes.Any(IsTargetEncoder(CodecSubtypes.AudioFormatFlac)))
-        {
-            return MediaEncodingProfile.CreateFlac(TranscodeQuality);
-        }
-
-        throw new NotImplementedException("尚未实现对目标编码器的支持。");
-
-        static Func<string, bool> IsTargetEncoder(string targetEncoderGuid)
-        {
-            return encoderGuid => encoderGuid == targetEncoderGuid;
-        }
     }
 
     private static async Task WriteTagsToFile(StorageFile musicFile, DownloadItem dlItem)
@@ -535,11 +564,13 @@ public static class DownloadService
         Uri coverUri = await FileCacheHelper.GetAlbumCoverUriAsync(albumDetail);
         try
         {
+            string defaultMsrName = "MSR".GetLocalized();
+
             List<SongInfo> songs = [.. albumDetail.Songs];
-            file.Tag.Performers = songDetail.Artists.Any() ? [.. songDetail.Artists] : ["MSR".GetLocalized()];
+            file.Tag.Performers = songDetail.Artists.Any() ? [.. songDetail.Artists] : [defaultMsrName];
             file.Tag.Title = songDetail.Name;
             file.Tag.Album = albumDetail.Name;
-            file.Tag.AlbumArtists = [songDetail.Artists.FirstOrDefault() ?? "MSR".GetLocalized()];
+            file.Tag.AlbumArtists = [songDetail.Artists.FirstOrDefault() ?? defaultMsrName];
             file.Tag.Track = (uint)songs.FindIndex(info => info.Cid == songDetail.Cid) + 1;
 
             TagLib.Picture picture;
@@ -593,88 +624,6 @@ public static class DownloadService
             ? 0d
             : (double)progress.BytesReceived / op.Progress.TotalBytesToReceive;
     }
-
-    //private static async Task TranscodeFile(IStorageFile sourceFile, IStorageFile destinationFile, AudioFormat format, AudioEncodingQuality quality, DownloadItem dlItem)
-    //{
-    //    CreateAudioGraphResult result = await AudioGraph.CreateAsync(new(Windows.Media.Render.AudioRenderCategory.Media));
-
-    //    if (result.Status != AudioGraphCreationStatus.Success)
-    //    {
-    //        throw new InvalidOperationException($"转码操作失败，原因：{result.Status}。");
-    //    }
-
-    //    using AudioGraph graph = result.Graph;
-    //    CreateAudioFileInputNodeResult inputResult = await graph.CreateFileInputNodeAsync(sourceFile);
-    //    if (inputResult.Status != AudioFileNodeCreationStatus.Success)
-    //    {
-    //        throw new InvalidOperationException($"转码操作失败，原因：{inputResult.Status}。");
-    //    }
-    //    using AudioFileInputNode inputNode = inputResult.FileInputNode;
-
-    //    MediaEncodingProfile profile = format switch
-    //    {
-    //        _ when quality is AudioEncodingQuality.Auto => throw new ArgumentOutOfRangeException(nameof(quality), "不能将音频质量设为 Auto。"),
-    //        AudioFormat.Mp3 => MediaEncodingProfile.CreateMp3(quality),
-    //        AudioFormat.Flac => MediaEncodingProfile.CreateFlac(quality),
-    //        _ => throw new NotImplementedException("尚未实现对指定编码器的支持。")
-    //    };
-    //    CreateAudioFileOutputNodeResult outputResult = await graph.CreateFileOutputNodeAsync(destinationFile, profile);
-    //    if (outputResult.Status != AudioFileNodeCreationStatus.Success)
-    //    {
-    //        throw new InvalidOperationException($"转码操作失败，原因：{outputResult.Status}。");
-    //    }
-
-    //    AudioFileOutputNode outputNode = outputResult.FileOutputNode;
-
-    //    inputNode.AddOutgoingConnection(outputNode);
-
-    //    TaskCompletionSource<bool> completionTask = new();
-    //    inputNode.FileCompleted += (sender, args) =>
-    //    {
-    //        sender.Stop();
-    //        completionTask.SetResult(true);
-    //    };
-
-    //    graph.Start();
-
-    //    await completionTask.Task;
-
-    //    TranscodeFailureReason outputFileResult = await outputNode.FinalizeAsync();
-    //    if (outputFileResult != TranscodeFailureReason.None)
-    //    {
-    //        throw new InvalidOperationException($"转码操作失败，原因：{outputFileResult}。");
-    //    }
-    //    graph.Stop();
-
-    //    //int bitrate = quality switch
-    //    //{
-    //    //    AudioEncodingQuality.Low => 96,
-    //    //    AudioEncodingQuality.Medium => 128,
-    //    //    AudioEncodingQuality.High or _ => 192,
-    //    //};
-
-    //    //if (format is AudioFormat.Flac)
-    //    //{
-    //    //    // TODO: add flac support
-    //    //    throw new NotImplementedException("Please wait...");
-    //    //}
-
-    //    //// TODO: 请解决源文件是 MP3 的情况。
-    //    //using Stream inputStream = await sourceFile.OpenStreamForReadAsync();
-    //    //using Stream outputStream = await destinationFile.OpenStreamForWriteAsync();
-
-    //    //inputStream.Seek(0, SeekOrigin.Begin);
-    //    //outputStream.Seek(0, SeekOrigin.Begin);
-
-    //    //using WaveFileReader reader = new(inputStream);
-    //    //using LameMP3FileWriter writer = new(outputStream, reader.WaveFormat, bitrate);
-    //    //long readerLength = reader.Length;
-    //    //writer.OnProgress += (writer, inputBytes, outputBytes, finished) =>
-    //    //{
-    //    //    dlItem.Progress = inputBytes / readerLength;
-    //    //};
-    //    //await reader.CopyToAsync(writer, 81920, dlItem.CancelToken.Token);
-    //}
 
     private static async Task TranscodeFile(IStorageFile sourceFile, IStorageFile destinationFile, AudioFormat format, AudioEncodingQuality quality, DownloadItem dlItem)
     {
